server/requirements.txt
Flask == 2.0.0
globus-sdk == 3.0.2
neo4j == 4.2.1
psycopg2-binary == 2.9.3
rdflib >= 6.0.0
requests == 2.26.0
urllib3 == 1.26.7